fix source_dir for tests

+++ cylc/flow/scheduler.py
@@ -42,7 +42,7 @@ from cylc.flow.cfgspec.glbl_cfg import glbl_cfg
 from cylc.flow.config import SuiteConfig
 from cylc.flow.cycling.loader import get_point
 from cylc.flow.exceptions import (
-    CylcError, SuiteConfigError, PlatformLookupError
+    CylcError, SuiteConfigError, PlatformLookupError, SuiteServiceFileError
 )
 import cylc.flow.flags
 from cylc.flow.host_select import select_suite_host
@@ -292,6 +292,16 @@ class Scheduler:
         # Check if flow has been installed
         if not suite_files.is_installed(self.suite_run_dir):
             suite_files.register(self.suite)
+        # Install
+        try:
+            suite_files.get_suite_source_dir(self.suite)
+        except SuiteServiceFileError:
+            # Source path is assumed to be the run directory
+            suite_files.register(
+                flow_name=self.suite,
+                source=get_workflow_run_dir(
+                    self.suite))
+
         make_suite_run_tree(self.suite)
 
         # Create ZMQ keys
@@ -306,14 +316,16 @@ class Scheduler:
         for sub_dir in ["python", os.path.join("lib", "python")]:
             # TODO - eventually drop the deprecated "python" sub-dir.
             suite_py = os.path.join(self.suite_dir, sub_dir)
-            if os.path.isdir(suite_py):
+            if (os.path.realpath(self.suite_dir) !=
+                    os.path.realpath(self.suite_run_dir) and
+                    os.path.isdir(suite_py)):
                 suite_run_py = os.path.join(self.suite_run_dir, sub_dir)
                 try:
                     rmtree(suite_run_py)
                 except OSError:
                     pass
                 copytree(suite_py, suite_run_py)
-            sys.path.append(os.path.join(self.suite_run_dir, sub_dir))
+            sys.path.append(os.path.join(self.suite_dir, sub_dir))
 
     async def initialise(self):
         """Initialise the components and sub-systems required to run the flow.
@@ -374,7 +386,7 @@ class Scheduler:
             proc_pool=self.proc_pool,
             suite_run_dir=self.suite_run_dir,
             suite_share_dir=self.suite_share_dir,
-            suite_source_dir=self.suite_run_dir
+            suite_source_dir=self.suite_dir
         )
 
         self.task_events_mgr = TaskEventsManager(
@@ -421,7 +433,9 @@ class Scheduler:
         for sub_dir in ["python", os.path.join("lib", "python")]:
             # TODO - eventually drop the deprecated "python" sub-dir.
             suite_py = os.path.join(self.suite_dir, sub_dir)
-            if os.path.isdir(suite_py):
+            if (os.path.realpath(self.suite_dir) !=
+                    os.path.realpath(self.suite_run_dir) and
+                    os.path.isdir(suite_py)):
                 suite_run_py = os.path.join(self.suite_run_dir, sub_dir)
                 try:
                     rmtree(suite_run_py)
